PythonProject5/requirements.txt
fastapi
uvicorn
requests
pydantic
python-dotenv
pytest
sqlalchemy
boto3
opencv-python-headless
twilio
python-multipart
pymysql
